requirements.txt
langchain
langchain-community
python-dotenv
tiktoken
openai
streamlit